+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v1.0.0-beta.91
+
+### instant-apps-export
+
+- Fix: popup load correctly and turn canvas to png, `extraContent` is removed when checked off, update CSS
+
 ## v1.0.0-beta.90
 
 ### instant-apps-interactive-legend